(PC-34944) refactor(venueMap): merge FilterBannerContainer and FilterCategoriesBannerContainer
