+++ src/features/venueMap/pages/VenueMap/VenueMapBase.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components/native'
 
 import { getSearchStackConfig } from 'features/navigation/SearchStackNavigator/helpers'
 import { useGoBack } from 'features/navigation/useGoBack'
-import { FilterBannerContainer } from 'features/venueMap/components/FilterBannerContainer/FilterBannerContainer'
+import { FilterCategoriesBannerContainer } from 'features/venueMap/components/FilterBannerContainer/FilterCategoriesBannerContainer'
 import { venuesFilterActions } from 'features/venueMap/store/venuesFilterStore'
 import { PageHeaderWithoutPlaceholder } from 'ui/components/headers/PageHeaderWithoutPlaceholder'
 
@@ -19,7 +19,7 @@ export const VenueMapBase: FunctionComponent<PropsWithChildren> = ({ children })
   return (
     <Container>
       <StyledHeader title="Carte des lieux" onGoBack={handleGoBack} />
-      <FilterBannerContainer />
+      <FilterCategoriesBannerContainer />
       {children}
     </Container>
   )
